# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "O2geeks Headless CMS",
    "description": "\n**Backend API for the O2geeks Branding System**\n\nThis API powers the decoupled content management backend. It provides robust RESTful endpoints for delivering and managing multi-format content.\n\n### Key Features:\n- **Authentication**: JWT-based secure access with Refresh Tokens.\n- **RBAC**: Strict Role-Based Access Control (Super Admin, Admin, Editor, User).\n- **Rate Limiting**: Distributed Redis-backed abuse protection.\n- **Media**: Dual-provider (Local + AWS S3) Multipart/Form-Data uploads.\n- **Engagement**: Zero-N+1 likes, comments, and favorites.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "List Users",
        "description": "List all users.",
        "operationId": "list_users_api_v1_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_UserOut_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User Management"
        ],
        "summary": "Create User",
        "description": "Create a new user.",
        "operationId": "create_user_api_v1_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Get User",
        "description": "Get a user by ID.",
        "operationId": "get_user_api_v1_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User Management"
        ],
        "summary": "Update User",
        "description": "Update a user's information.",
        "operationId": "update_user_api_v1_users__user_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User Management"
        ],
        "summary": "Deactivate User",
        "description": "Deactivate a user.",
        "operationId": "deactivate_user_api_v1_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_UserOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[UserOut]"
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "title": "UserCreate",
        "description": "Request body for admin creating a new user.\n\nAllowed roles:\n- super_admin\n- admin\n- editor\n- user\n- viewer"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "created_at",
          "updated_at",
          "permissions"
        ],
        "title": "UserOut",
        "description": "Response model (never exposes password)."
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "super_admin",
          "admin",
          "editor",
          "user",
          "viewer"
        ],
        "title": "UserRole"
      },
      "UserUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Admin updating another user."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}